% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%>\%}
\alias{\%>\%}
\alias{\%<>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs

lhs \%<>\% rhs
}
\value{
The left-hand side value is passed to the right-hand side function.

The left-hand side value is modified by the right-hand side function and reassigned to the left-hand side.
}
\description{
This operator allows for chaining commands in a more readable way.

This operator allows for chaining commands in a more readable way, while also updating the left-hand side value.
}
\keyword{internal}
