% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_assessment_id.R
\name{rl_assessment_id}
\alias{rl_assessment_id}
\title{IUCN Red List assessment}
\usage{
rl_assessment_id(assessment_id = 1425064)
}
\arguments{
\item{assessment_id}{Assessment ID}
}
\value{
A tibble where each column represents a unique API response JSON key for the supplied \code{assessment_id}.
The columns include key information about the Red List assessment, such as taxon details, category, year, and other relevant metadata.
}
\description{
Retrieves an assessment
}
\examples{
\dontrun{
rl_assessment_id(1425064)
}

}
