% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_get_count.R
\name{rl_get_count}
\alias{rl_get_count}
\title{Get Total Number of Species Assessed by IUCN for a Given Country}
\usage{
rl_get_count(country_code = "BJ")
}
\arguments{
\item{country_code}{A character string specifying the ISO 3166-1 alpha-2
country code for which the species count is to be retrieved. Defaults to \code{"BJ"} (Benin).}
}
\value{
A numeric value representing the total number of species assessed for the specified country.
}
\description{
This function retrieves the total number of species assessed by the IUCN Red
List for a specified country.
}
\details{
The function queries the IUCN Red List API to obtain the total count of species
assessed for the specified country. It requires your API key.
}
\examples{
# Example usage
\dontrun{
# Set your IUCN API key as an environment variable
rl_set_api("your_api_key_here")

# Get the total species count for Benin (default)
total_species <- rl_get_count()
print(total_species)

# Get the total species count for India
total_species_india <- rl_get_count(country_code = "IN")
print(total_species_india)
}

}
