% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_green_status.R
\name{rl_green_status}
\alias{rl_green_status}
\title{Green status}
\usage{
rl_green_status()
}
\value{
A tibble (class \code{tbl_df}, \code{tbl}, \code{data.frame}) where each column represents a unique API response JSON key.
The columns include key information about the Green Status assessment, such as year, weights, justification, and
other relevant metadata.
}
\description{
Retrieve all IUCN Green Status assessments.
}
\examples{
\dontrun{
rl_green_status()
}
}
