% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_possibly_extinct.R
\name{rl_possibly_extinct}
\alias{rl_possibly_extinct}
\title{IUCN Red List possibly extinct taxa}
\usage{
rl_possibly_extinct()
}
\value{
A tibble (class \code{tbl_df}, \code{tbl}, \code{data.frame}) where each column represents a unique API response JSON key.
Columns include year, latest, possibly extinct, possibly extinct in the wild, sis taxon id, url, taxon scientific name,
red list category, assessment id, scopes description, and scopes code
}
\description{
Retrieve species assessments flagged as possibly extinct.
Returns all latest global assessments for taxa that are possibly extinct.
}
\examples{
\dontrun{
# Get all possibly extinct species
rl_possibly_extinct()
}
}
