% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_set_api.R
\name{rl_set_api}
\alias{rl_set_api}
\title{Set the IUCN Red List API key}
\usage{
rl_set_api(api_key)
}
\arguments{
\item{api_key}{Character. The API key provided by the IUCN Red List to
authenticate requests, obtainable at \href{https://api.iucnredlist.org/users/sign_in}{IUCN Red List API website}.}
}
\value{
Invisibly returns \code{NULL} after setting the API key.
}
\description{
The function provide steps to set the IUCN Red List API key.
}
\examples{
\dontrun{
# Set the API key for the IUCN Red List
rl_set_api("your_api_key")
}

}
