% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rl_total_records}
\alias{rl_total_records}
\title{Get total number of species assessed by IUCN}
\usage{
rl_total_records(url)
}
\arguments{
\item{url}{A character string specifying the query url}
}
\value{
A numeric value representing the total number of species assessed.
}
\description{
This function retrieves the total number of species assessed by the IUCN Red
List
}
