% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_assessment_id.R
\name{rl_assessment_id}
\alias{rl_assessment_id}
\title{IUCN Red List assessment}
\usage{
rl_assessment_id(assessment_id = 1425064)
}
\arguments{
\item{assessment_id}{Appeasement ID}
}
\value{
Returns assessment data for a supplied assessment_id.
}
\description{
Retrieves an assessment
}
\examples{
\dontrun{
rl_assessment_id(1425064)
}

}
