% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_country.R
\name{rl_country}
\alias{rl_country}
\title{Retrieve IUCN Red List Assessments by Country}
\usage{
rl_country(country_code = "BJ", year = NULL, page = NULL, verbose = TRUE)
}
\arguments{
\item{country_code}{Character. The ISO 3166-1 alpha-2 country code (e.g., "BJ" for Benin). Defaults to "BJ".}

\item{year}{Numeric or NULL. The year of the assessments to retrieve. If NULL, all years are retrieved.}

\item{page}{Numeric, character, or NULL. Specifies the page number(s) of data to retrieve.
\itemize{
\item If NULL, retrieves the first page.
\item If a single numeric value, retrieves that specific page.
\item If "all", retrieves all pages available.
\item If a numeric vector, retrieves multiple specific pages.
}}

\item{verbose}{Logical. If TRUE, prints progress messages when multiple pages are requested. Defaults to TRUE.}
}
\value{
A data frame containing the species assessment data, excluding columns \code{scopes} and \code{code_type}.
}
\description{
Fetches species assessments from the IUCN Red List API for a specified country and year.
This function returns data on threatened species assessed within the specified country,
filtered by year and/or page if specified.
}
\details{
The function first checks the Red List API authorization, then constructs an API request based on the
specified parameters (\code{country_code}, \code{year}, and \code{page}). If \code{page} is set to "all", it automatically
retrieves all available pages of data by calling the \code{\link[=rl_get_count]{rl_get_count()}} function to determine the total
number of pages required.

The retrieved data frame contains the following fields:
\itemize{
\item \code{taxonid}: Taxon ID of the species.
\item \code{scientific_name}: Scientific name of the species.
\item \code{year_published}: The year the assessment was published.
\item and other relevant assessment details.
}
}
\note{
Requires an IUCN Red List API token to be set as an environment variable under \code{redlist_api} with \code{\link[=rl_set_api]{rl_set_api()}} function.
For reliable functioning, the helper function \code{\link[=rl_check_api]{rl_check_api()}} could be called to validate the API token.
}
\examples{
\dontrun{
  # Retrieve assessments for Benin (country code "BJ") for the year 2020
  rl_country("BJ", year = 2020)

  # Retrieve all assessments for Brazil (country code "BR")
  rl_country("BR", page = "all")

  # Retrieve assessments for Canada (country code "CA") on specific pages
  rl_country("CA", page = c(1, 2))
}

}
