% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openapi.R
\name{rl_statistics}
\alias{rl_statistics}
\title{IUCN Red List statistics}
\usage{
rl_statistics()
}
\value{
A tibble containing the count of assessed species.
}
\description{
Retrieve count of species with assessments.
This endpoint returns the total number of assessed species on the IUCN Red List.
}
\examples{
\dontrun{
# Get total count of assessed species
rl_statistics()
}
}
