% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_version.R
\name{rl_version}
\alias{rl_version}
\title{IUCN Red List and API version}
\usage{
rl_version()
}
\description{
Returns the current version of the IUCN Red List of Threatened Species and API
}
\examples{
\dontrun{
rl_version()
}
}
